==> Analysis/01 Data Conversion - automated for processing schema Ver 03 input.ipynb <==
# Convert and upload the json input data to the Sqlite database
All the real work happens in the Civ6CitySettlementJsonLoader python module in the Analysis folder of this project.

The Civ6CitySettlementJsonLoader.Loader class is initialised with the json file to import. It follows the following high level plan:
1. Checks if the jsonDataFilename exists and retrieves the "jsonSchemaVersion" number.
2. Validates the jsonDataFile content:
       2.1 against the jsonSchemaVersion, i.e. tests the json input is in the expected format.
       2.2 by loading it into an InMemory database to ensure no load errors
3. Then opens the actual sqlite database connection and uploads the data

import Civ6CitySettlementJsonLoader as ccjl

Uncomment the following cell and then run the notebook to upload the json files to the database. It is commented out by default to avoid loading the same datafile multiple times. At the moment there is no sensible way to prevent data being uploaded multiple times.

It is possible to remove the duplicated data, but it is a pain. Handle this step with care!

#FilesToLoad = ['RawData/20181129_01.json', 'RawData/20181129_02.json', 'RawData/20181129_03.json']
#FilesToLoad = ['RawData/20181207_01.json']
#FilesToLoad = ['RawData/20181208_01.json']
#FilesToLoad = ['RawData/20181208_02.json', 'RawData/20181208_03.json']
#FilesToLoad = ['RawData/20181209_01.json']
#FilesToLoad = ['RawData/20181209_02.json', 'RawData/20181209_03.json']
#FilesToLoad = ['RawData/20181210_01.json']
#FilesToLoad = ['RawData/20181211_01.json']
#FilesToLoad = ['RawData/20181213_01.json', 'RawData/20181213_02.json', 'RawData/20181213_03.json']
#FilesToLoad = ['RawData/20181213_04.json']
#FilesToLoad = ['RawData/20181214_01.json']
#FilesToLoad = ['RawData/20181214_02.json']
#FilesToLoad = ['RawData/20181214_03.json']
FilesToLoad = ['RawData/20181214_04.json', 'RawData/20181214_05.json', 'RawData/20181214_06.json']

for file in FilesToLoad:
    loader = ccjl.Loader(file)
    if (loader.hasInitialsedCorrectly == True):
        if (loader.jsonDataPassesValidation()):
            db = loader.connectDb('Database/Civ6CitySettledData.db')
            if (db is not None):
                isSuccessful = loader.uploadToDatabase(db)
                if(isSuccessful):
                    print("\nCompleted upload successfully!")
                else:
                    print("\nFailed upload, abort!")
            else:
                print("\nFailed database connection, abort!")
        else:
            print("\nFailed validation, abort!")
    else:
        print("\nFailed initilisation, abort!")


#print(loader)
#print("\nCurrent functions and variables:")
#for row in dir(loader):
#    if not(row.startswith("__")):
#        print("\t", row)
