==> Analysis/03 Data Exploration - City Plots Settled.ipynb <==
# Explore the plots, when settling a city, measurements

The database contains all the plots in the first two rings from the city centre, that is:
- city centre is in ring 0 and contains 1 plot,
- adjacent to city tiles are ring 1 and contains 6 plots,
- those 2 tiles away are ring 2 and contains 12 plots.

That means we capture the 19 nearest tiles' plot information.

Data captured per tile includes:
- owner of plot at the time the city was settled. (Most civ's start with 7/19 tiles)
- recordedCityId is an identifier of all the tiles that "could" belong to a city when settled
- terrain information
- features information
- resource, resourceCount, and resourceType information
- workers on plot count
- district on plot - this is unlikely to be anything other than 'City Centre'
- hasRiver - does the plot have an adjacent river
- isWater - is the plot water
- isLake - is the plot part of a lake
- isCity - only true once per city...

The purpose of this workbook is to investigate how to present this information as input to the ML model.

I purposely did not capture the "yields" to make the Machine Learning Classification challenge more realistic. Also, there is no relation between any of the database Id values and the in-game Ids

import sqlite3
import matplotlib.pyplot as plt
import pandas as pd
#pd.set_option('display.height', 1000)
pd.set_option('display.max_rows', 200)
pd.set_option('display.max_columns', 10)
pd.set_option('display.width', 700)
pd.set_option('display.max_colwidth', 100)

cnx = sqlite3.connect('Database/Civ6CitySettledData.db')
cur = cnx.cursor()
#print(cnx)
#print(cur)

## Retrieve the city plots as settled data
cityPlotsSettled contains the city plot, adjacent to city plots, and 2 tiles from city plots information at the time the city was settled. The idea is to look at this information and "predict" the yield performance of the city.

The primary identity for this database is the cityId. That is, this database (and experiment) is designed to look at city growth over time. The same cityName, leaderName etc. can occur in multiple games so when we add the recorded information for the game we create new cityIds.

# After experimentation I decided to use all the plots within 2 tiles from the city centre.
sqlSelect = 'SELECT * FROM cityPlotsSettled'

# Other options I experimented with are:
#sqlSelect = 'SELECT * FROM cityPlotsSettled WHERE ownerCityId <> "None"'
#sqlSelect = 'SELECT * FROM cityPlotsSettled WHERE ring == 0'
#sqlSelect = 'SELECT * FROM cityPlotsSettled WHERE ring == 1'
#sqlSelect = 'SELECT * FROM cityPlotsSettled WHERE ring == 2'

cityPs = pd.read_sql_query(sqlSelect, cnx)
print(cityPs.shape)
print(cityPs.dtypes)

## Plot terrain counts

df = cityPs[['plotId', 'terrain']].groupby('terrain').count().sort_values(by='plotId', ascending=False)
df.rename(columns={'plotId':'Count'}, inplace=True)
print(df)

## Plot features counts
City doesn't have features as they are "removed" when settled

df = cityPs[['plotId', 'feature']].groupby('feature').count().sort_values(by='plotId', ascending=False)
df.rename(columns={'plotId':'Count'}, inplace=True)
print(df)

## Plot resource counts

df = cityPs[['plotId', 'resource']].groupby('resource').count().sort_values(by='plotId', ascending=False)
#print(len(cityPs[['plotId', 'resource']].groupby('resource').count().sort_values(by='plotId', ascending=False)))
df.rename(columns={'plotId':'Count'}, inplace=True)
print(df)

#print(cityPs[['plotId', 'resourceType']].groupby('resourceType').count().sort_values(by='plotId', ascending=False))

#print(cityPs[['plotId', 'district']].groupby('district').count().sort_values(by='plotId', ascending=False))

df = cityPs[['plotId', 'hasRiver']].groupby('hasRiver').count().sort_values(by='plotId', ascending=False)
df.rename(columns={'plotId':'Count'}, inplace=True)
print(df)

#print(cityPs[['plotId', 'resourceCount']].groupby('resourceCount').count().sort_values(by='plotId', ascending=False))

#cityPs[cityPs['resourceType'] == 'Strategic'].resource.unique()
#cityPs[cityPs['resourceType'] == 'Luxury'].resource.unique()
#cityPs[cityPs['resourceType'] == 'Bonus'].resource.unique()

#cols = list(cityPs['category'].unique())
#print(type(cols))
#print(len(cols))
#print(cols)

cur.close()
cnx.close()
